Refactored keyword visitor
