+++ src/robotframework_find_unused/visitors/robot/keyword_visitor/__init__.py
@@ -1,0 +1,7 @@
+"""
+Robot Framework AST visitor for keywords
+"""
+
+from .keyword_visitor import RobotVisitorKeywords
+
+__all__ = ["RobotVisitorKeywords"]
